// packages/vincent-rest-api/src/generated/api/models/IToolDef.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
 
export type IToolDef = {
  /**
   * Tool package name
   */
  packageName: string;
  /**
   * Tool title
   */
  toolTitle?: string;
  /**
   * Unique composite identifier
   */
  identity: string;
  /**
   * Author wallet address
   */
  authorWalletAddress: string;
  /**
   * Tool description
   */
  description: string;
  /**
   * Active version of the tool
   */
  activeVersion: string;
};